ref: minor refactor
