+++ internal/state/utils.go
@@ -1,28 +1,10 @@
-package internal
+package state
 
 import (
 	"encoding/json"
 	"fmt"
 	"os"
 )
-
-type PersistentState struct{
-  CurrentTerm int `json:"current_term"`
-  VotedFor int `json:"voted_for"`
-}
-
-type LeaderState struct{
-  nextIndex map[string]int
-  matchIndex map[string]int
-}
-
-type State struct{
-  Id int 
-  CommitIndex int
-  LastAppliedIndex int
-  PersistentState PersistentState
-  Log []string
-}
 
 
 func (s *State) AppendLog(entry string) error {
